architecture: split obs/signoz.go into focused modules
obs/client.go already existed implicitly in signoz.go. Now explicit:

- obs/signoz.go — SigNozClient, HTTP transport, cache, v3 types, query
- obs/extract.go — BuildServices, BuildHosts, LatestByDeployment,
  SparkByDeployment, LatestByNode (pure extraction functions)
- obs/topology.go — CategoryForNamespace, DiscoverNodeNames,
  SortServices, LabelStr (category mapping and discovery)
- obs/models.go — Service, Host, LiveData, Client interface (types only)

Deprecated aliases kept for backward compat: discoverNodeNames,
labelStr, sortServices. The live template calls these from signoz.go's
v3 types which remain there.

Extraction functions are now testable with fixture v3Series data
without involving the HTTP client.
